+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/User.java
@@ -11,8 +11,7 @@ import java.util.Set;
 
 public class User {
 
-
-   private Long id;
+   private Integer id;
 
    @Size(min = 3, max = 50, message = "Username must be in between 3 and 50 characters.")
    private String username;
@@ -25,18 +24,18 @@ public class User {
 
    public User() { }
 
-   public User(Long id, String username, String password, String authorities) {
+   public User(Integer id, String username, String password, String authorities) {
       this.id = id;
       this.username = username;
       this.password = password;
       this.activated = true;
    }
 
-   public Long getId() {
+   public Integer getId() {
       return id;
    }
 
-   public void setId(Long id) {
+   public void setId(Integer id) {
       this.id = id;
    }
